feat: tasks are now linked to the user who created them

+++ src/main/java/com/antonio/taskmanager/service/TaskService.java
@@ -1,11 +1,13 @@
 package com.antonio.taskmanager.service;
 
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import com.antonio.taskmanager.dto.*;
 import com.antonio.taskmanager.repository.TaskRepository;
 
 import jakarta.validation.Valid;
+import lombok.AllArgsConstructor;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -16,24 +18,26 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.antonio.taskmanager.entity.Task;
+import com.antonio.taskmanager.entity.User;
+import com.antonio.taskmanager.enums.Role;
 import com.antonio.taskmanager.mapper.TaskMapper;
 
 @Service
+@AllArgsConstructor
 public class TaskService {
     private final TaskRepository taskRepository;
     private final TaskMapper taskMapper;
+    private final AuthService authService;
     private static final Logger logger = LoggerFactory.getLogger(TaskService.class);
-
-    public TaskService(TaskRepository taskRepository, TaskMapper taskMapper) {
-        this.taskRepository = taskRepository;
-        this.taskMapper = taskMapper;
-    }
 
     public TaskResponseDTO createTask(@Valid TaskRequestDTO dto) {
         logger.info("Creating a new task with title: {}", dto.getTitle());
 
         Task task = taskMapper.toEntity(dto);
         logger.debug("Entity created: {}", task);
+
+        User currentUser = authService.getCurrentUser();
+        task.setUser(currentUser);
 
         Task savedTask = taskRepository.save(task);
         logger.info("Task saved with ID: {}", savedTask.getId());
@@ -42,7 +46,16 @@ public class TaskService {
     }
 
     public List<TaskResponseDTO> getAllTasks() {
-        List<Task> tasks = taskRepository.findAll();
+        User currentUser = authService.getCurrentUser();
+
+        List<Task> tasks;
+        if (currentUser.getRole() == Role.ADMIN) { // admin can see all tasks
+            tasks = taskRepository.findAll();
+        } else { // regular user can only see their own tasks
+            tasks = taskRepository.findAllByUser(currentUser);
+            logger.info("Tasks found for user: {}", currentUser.getId());
+        }
+
         return tasks.stream()
                 .map(taskMapper::toDTO)
                 .collect(Collectors.toList());
@@ -52,6 +65,13 @@ public class TaskService {
         Task task = taskRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Task not found with ID: " + id));
         logger.info("Task found with ID: {}", task.getId());
+
+        User currentUser = authService.getCurrentUser(); // get the current user
+        if (!task.getUser().getId().equals(currentUser.getId())) { // this line checks if the task belongs to the
+                                                                   // current user
+            throw new RuntimeException("You do not have permission to access this task");
+        }
+
         return taskMapper.toDTO(task);
     }
 
@@ -59,6 +79,11 @@ public class TaskService {
         Task task = taskRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Task not found with ID: " + id));
         logger.info("Task found with ID: {}", task.getId());
+
+        User currentUser = authService.getCurrentUser();
+        if (!task.getUser().getId().equals(currentUser.getId())) {
+            throw new AccessDeniedException("You do not have access to this task");
+        }
 
         if (dto.getTitle() != null) {
             task.setTitle(dto.getTitle());
@@ -86,6 +111,11 @@ public class TaskService {
                 .orElseThrow(() -> new RuntimeException("Task not found with ID: " + id));
         logger.info("Task found with ID: {}", task.getId());
 
+        User currentUser = authService.getCurrentUser();
+        if (!task.getUser().getId().equals(currentUser.getId())) {
+            throw new AccessDeniedException("You do not have access to this task");
+        }
+
         taskRepository.delete(task);
         logger.info("Task deleted with ID: {}", task.getId());
     }
@@ -94,6 +124,11 @@ public class TaskService {
         Task task = taskRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Task not found with ID: " + id));
         logger.info("Task found with ID: {}", task.getId());
+
+        User currentUser = authService.getCurrentUser();
+        if (!task.getUser().getId().equals(currentUser.getId())) {
+            throw new AccessDeniedException("You do not have access to this task");
+        }
 
         if (dto.getTitle() != null)
             task.setTitle(dto.getTitle());
